backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
pydantic==2.5.3
